python/requirements.txt
fastapi
python-multipart
uvicorn[standard]
chromadb
sentence-transformers
transformers
torch
networkx
pywebview
httpx
pytest
pytest-asyncio
